+++ CmdLine.cc
@@ -500,15 +500,13 @@ string CmdLine::command_line() const {
 
 
 bool CmdLine::Error::_do_printout = true;
-CmdLine::Error::Error(const std::ostringstream & ostr) 
-  : std::runtime_error("CmdLine Error: " + ostr.str()) {
+CmdLine::Error::Error(const std::ostringstream & ostr) : CmdLine::Error(ostr.str()) {}
+
+  CmdLine::Error::Error(const std::string & str) 
+  : std::runtime_error(str) {
   _message = what();
-  if (_do_printout) cerr <<  _message << endl;;
-}
-CmdLine::Error::Error(const std::string & str) 
-  : std::runtime_error("CmdLine Error: " + str) {
-  _message = what();
-  if (_do_printout) cerr << _message << endl;;
+  if (_do_printout) cerr << tc::red << tc::bold 
+                         << "CmdLine Error: " << tc::nobold << _message << tc::reset << endl;
 }
 
 string CmdLine::current_path() const {
@@ -1083,3 +1081,57 @@ template<> bool CmdLine_string_to_value<bool>(const std::string & str) {
   if (lcstr == "0" || lcstr == "no" || lcstr == "off" || lcstr == "false" || lcstr == ".false.") return false;
   throw CmdLine::ConversionFailure(str);
 }
+
+
+// all the terminal control strings
+std::string CmdLine::tc::red = "\033[31m";
+std::string CmdLine::tc::grn = "\033[1;32m";
+std::string CmdLine::tc::yel = "\033[1;33m";
+std::string CmdLine::tc::blu = "\033[1;34m";
+std::string CmdLine::tc::mag = "\033[1;35m";
+std::string CmdLine::tc::cyn = "\033[1;36m";
+std::string CmdLine::tc::wht = "\033[1;37m";
+std::string CmdLine::tc::blk = "\033[1;30m";
+std::string CmdLine::tc::gry = "\033[1;90m";
+std::string CmdLine::tc::org = "\033[1;91m";
+
+std::string CmdLine::tc::red_bg = "\033[41m";
+std::string CmdLine::tc::grn_bg = "\033[42m";
+std::string CmdLine::tc::yel_bg = "\033[43m";
+std::string CmdLine::tc::blu_bg = "\033[44m";
+std::string CmdLine::tc::mag_bg = "\033[45m";
+std::string CmdLine::tc::cyn_bg = "\033[46m";
+std::string CmdLine::tc::wht_bg = "\033[47m";
+std::string CmdLine::tc::blk_bg = "\033[40m";
+std::string CmdLine::tc::gry_bg = "\033[100m";
+std::string CmdLine::tc::org_bg = "\033[101m";
+
+// versions with full names for those colors that have >3 letters
+// both foreground and background
+std::string CmdLine::tc::yellow = "\033[1;33m";
+std::string CmdLine::tc::blue = "\033[1;34m";
+std::string CmdLine::tc::magenta = "\033[1;35m";
+std::string CmdLine::tc::cyan = "\033[1;36m";
+std::string CmdLine::tc::white = "\033[1;37m";
+std::string CmdLine::tc::black = "\033[1;30m";
+std::string CmdLine::tc::gray = "\033[1;90m";
+std::string CmdLine::tc::orange = "\033[1;91m";
+
+std::string CmdLine::tc::yellow_bg = "\033[43m";
+std::string CmdLine::tc::blue_bg = "\033[44m";
+std::string CmdLine::tc::magenta_bg = "\033[45m";
+std::string CmdLine::tc::cyan_bg = "\033[46m";
+std::string CmdLine::tc::white_bg = "\033[47m";
+std::string CmdLine::tc::black_bg = "\033[40m";
+std::string CmdLine::tc::gray_bg = "\033[100m";
+std::string CmdLine::tc::orange_bg = "\033[101m";
+
+
+std::string CmdLine::tc::bold = "\033[1m";
+std::string CmdLine::tc::nobold = "\033[22m";
+std::string CmdLine::tc::underline = "\033[4m";
+std::string CmdLine::tc::reverse = "\033[7m";
+std::string CmdLine::tc::reset = "\033[0m";
+std::string CmdLine::tc::clear = "\033[2J\033[H"; ///< clear screen and move cursor to home
+std::string CmdLine::tc::clear_screen = "\033[2J";
+std::string CmdLine::tc::clear_line = "\033[2K\r";